Move filter tests to their own file and fix subsequent local import issues by setting up a conftest file
